CADY home task
Task:
 The attached files contain text obtained (in theory) from components datasheet.
 Write a system in Python with the following capabilities:
o Reading the content of all the given components&#39; text files.
o Parsing and storing for each file component the operating voltage range
(i.e. the voltages values in which the component can function; e.g. from 5V
to 12V) and the operating temperature range (i.e. the temperatures in
which the component can function (e.g. from -40C to 85C).
o All files contain at least one range for the each of the items. If a file has
more than one range for an item check if they are identical. If they are, use
it as the parsed range. If not, write None as the range.
o Given specific operating conditions return all the components that can
operate under these conditions. The conditions are given as two floats, one
for the voltage and one for the temperature, and they are assumed to be in
the standard units (e.g. 5 for 5V and -20 for -20C).
This is the interface with which the system is used.